gemini_backend_deployment_for_shapeshift/basic_chat.py
# Install the Google AI Python SDK
# $ pip install google-generativeai
from dotenv import load_dotenv
import os
import google.generativeai as genai



def generate_basic_chat(user_input):
    
    load_dotenv()
    # Configure the API key
    api_key = os.getenv("GEMINI_API_KEY")
    genai.configure(api_key=api_key)

    # Create the model configuration
    # Create the model
    generation_config = {
    "temperature": 1,
    "top_p": 0.95,
    "top_k": 40,
    "max_output_tokens": 8192,
    "response_mime_type": "text/plain",
    }

    model = genai.GenerativeModel(
    model_name="gemini-1.5-flash-002",
    generation_config=generation_config,
    # safety_settings = Adjust safety settings
    # See https://ai.google.dev/gemini-api/docs/safety-settings
    )

    chat_session = model.start_chat(
    history=[
    ]
    )

    response = chat_session.send_message(user_input)

    return response.text
